engine: fixing pki handling for successive executions
Ensuring the engine container can run a second time with the
PKI created by the first run of engine-setup:
on the first run engine-setup creates the oVirt PKI writing it
on a persistent volume. engine-setup requires a template there
but the initial persistent volume is by default empy and so we
have to copy the template there on the first execution and avoid
overwriting it on successive ones.
We have also to force the ownership of the PKI volume since
in kubernetes the ownership of mounted volumes is always
set at root while the engine requires ovirt:ovirt. See
https://github.com/kubernetes/kubernetes/issues/2630

+++ image-specifications/engine/entrypoint.sh
@@ -13,9 +13,13 @@ echo OVESETUP_CONFIG/adminPassword=str:$OVIRT_PASSWORD >> answers.conf
 echo OVESETUP_PKI/organization=str:$OVIRT_PKI_ORGANIZATION >> answers.conf
 echo OVESETUP_CONFIG/adminUserId=str:$OVIRT_ADMIN_UID >> answers.conf
 
-# Copy pki template files into original template location.
-# Mounts on kubernetes hide the original files in the image.
-cp -a /etc/pki/ovirt-engine.tmpl/* /etc/pki/ovirt-engine/
+# On the first run, copy pki template files into original
+# template location because kubernetes mounts hide
+# the original files in the image.
+if [ ! -f /etc/pki/ovirt-engine/serial.txt ]; then
+    cp -a /etc/pki/ovirt-engine.tmpl/* /etc/pki/ovirt-engine/
+fi
+chown ovirt:ovirt /etc/pki/ovirt-engine
 
 # Wait for postgres
 dockerize -wait tcp://${POSTGRES_HOST}:${POSTGRES_PORT} -timeout 1m
